KTOR-4525 Add PartialContent Logging (#3199)
